+++ source/Public/Workspace/Remove-FabricWorkspace.ps1
@@ -26,6 +26,7 @@ function Remove-FabricWorkspace {
     param (
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
+        [Alias("Id")]
         [guid]$WorkspaceId
     )