+++ source/Public/Workspace/Remove-FabricWorkspaceIdentity.ps1
@@ -27,6 +27,7 @@ function Remove-FabricWorkspaceIdentity
     param (
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
+        [Alias("Id")]
         [guid]$WorkspaceId
     )