+++ source/Public/Workspace/Remove-FabricWorkspaceRoleAssignment.ps1
@@ -30,6 +30,7 @@ function Remove-FabricWorkspaceRoleAssignment
     param (
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
+        [Alias("Id")]
         [guid]$WorkspaceId,
 
         [Parameter(Mandatory = $true)]