+++ source/Public/Workspace/Update-FabricWorkspaceRoleAssignment.ps1
@@ -37,6 +37,7 @@ function Update-FabricWorkspaceRoleAssignment
     param (
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
+        [Alias("Id")]
         [guid]$WorkspaceId,
 
         [Parameter(Mandatory = $true)]